A lot of content
Working on getting the bulk of it in place, then I'll go back to edit and find some images to throw in. If there's anything you think needs changing let me know, but it should be pretty straightforward, and I'll be consulting over the assets to use for images before I put any in.
Also, for your entertainment:
Ever seen a rabbit on a trampoline? It's a hare raising experience!